requirements.txt
streamlit==1.8.1
Pillow
matplotlib
tensorflow==2.9.2
keras
numpy
pandas
